+++ mitra_api/src/mastodon_api/oauth/types.rs
@@ -24,6 +24,7 @@ pub struct TokenRequest {
 
     // Required if grant type is "authorization_code"
     pub code: Option<String>,
+    #[allow(dead_code)]
     pub redirect_uri: Option<String>,
     pub client_id: Option<Uuid>,