+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Added `timelines_access` property to `/api/v2/instance` response ([#191](https://codeberg.org/silverpill/mitra/pulls/191)).
 
+### Changed
+
+- Write client ID to log every time an access token is generated.
+
 ## [4.13.1] - 2025-11-19
 
 ### Changed